Fix pixel function selection
A "safe" pixel plotting function is used if the thing being
plotted goes out the screen. Previously the code which decides
which to use assumed that the first bounding box coordinate was
nearer the origin than the second bounding box coordinate.

Now we don't care what order they're in, making the library
easier to use.